requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
coverage==7.2.1
Django==2.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
et-xmlfile==1.1.0
idna==3.4
openpyxl==3.1.2
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==3.12.3
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2022.7.1
requests==2.28.2
sqlparse==0.2.4
tomli==2.0.1
urllib3==1.26.14
requests-toolbelt==0.9.1
pytest==7.2.2
xhtml2pdf==0.2.11